requirements.txt
torch
torchvision
matplotlib
numpy
pandas
Pillow
scikit-learn
torch_geometric
timm
transformers
ftfy
nbconvert